Add DAG support to tasks.yml

+++ src/stun/random_utils.cpp
@@ -17,7 +17,7 @@ int16 RandomUtils::uniform_int(int16 max) noexcept
 {
     Expects(max >= 0);
     using param_type = decltype(uniform_int_)::param_type;
-    uniform_int_.param(param_type{0, static_cast<int>(max)});
+    uniform_int_.param(param_type{0, max});
     return uniform_int_(generator);
 }
 } // namespace angonoka::stun
